backend/requirements.txt
fastapi>=0.75.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
starlette>=0.17.1
Pillow>=6.1.0
opencv-python>=4.8.1.78
numpy>=1.26.2
dlib>=19.24.2
mtcnn>=0.1.1